chore(dup): Reuse insert queries from DataAccess
Move functions for insterting device values
into DataAccess so it can be reused in DUP
and AppEngine
