+++ apps/astarte_appengine_api/lib/astarte_appengine_api/device/queries.ex
@@ -30,7 +30,6 @@ defmodule Astarte.AppEngine.API.Device.Queries do
   import Ecto.Query
 
   alias Astarte.AppEngine.API.Config
-  alias Astarte.AppEngine.API.DateTime, as: DateTimeMs
   alias Astarte.AppEngine.API.Device.DevicesList
   alias Astarte.AppEngine.API.Device.DeviceStatus
   alias Astarte.AppEngine.API.Device.InterfaceInfo
@@ -222,255 +221,6 @@ defmodule Astarte.AppEngine.API.Device.Queries do
     Repo.insert!(value, opts)
 
     :ok
-  end
-
-  # TODO Copy&pasted from data updater plant, make it a library
-  def insert_value_into_db(
-        realm_name,
-        device_id,
-        %InterfaceDescriptor{storage_type: :multi_interface_individual_properties_dbtable} =
-          interface_descriptor,
-        %Endpoint{allow_unset: true} = endpoint,
-        path,
-        nil,
-        _timestamp,
-        _opts
-      ) do
-    # TODO: :reception_timestamp_submillis is just a place holder right now
-    %InterfaceDescriptor{interface_id: interface_id, storage: storage} = interface_descriptor
-    %Endpoint{endpoint_id: endpoint_id} = endpoint
-    keyspace_name = Realm.keyspace_name(realm_name)
-
-    delete_match =
-      from v in storage,
-        prefix: ^keyspace_name,
-        where:
-          v.device_id == ^device_id and v.interface_id == ^interface_id and
-            v.endpoint_id == ^endpoint_id and v.path == ^path
-
-    {c, _} = Repo.delete_all(delete_match, consistency: Consistency.device_info(:write))
-
-    if c == 0 do
-      _ =
-        Logger.warning(
-          "Could not unset value for #{Device.encode_device_id(device_id)} in #{storage} or there was no data",
-          realm: "realm",
-          tag: "cant_unset"
-        )
-    end
-
-    :ok
-  end
-
-  def insert_value_into_db(
-        _realm_name,
-        _device_id,
-        %InterfaceDescriptor{storage_type: :multi_interface_individual_properties_dbtable} =
-          _interface_descriptor,
-        _endpoint,
-        _path,
-        nil,
-        _timestamp,
-        _opts
-      ) do
-    _ =
-      Logger.warning("Tried to unset value on allow_unset=false mapping.",
-        tag: "unset_not_allowed"
-      )
-
-    {:error, :unset_not_allowed}
-  end
-
-  # TODO Copy&pasted from data updater plant, make it a library
-  def insert_value_into_db(
-        realm_name,
-        device_id,
-        %InterfaceDescriptor{storage_type: :multi_interface_individual_properties_dbtable} =
-          interface_descriptor,
-        %Endpoint{} = endpoint,
-        path,
-        value,
-        timestamp,
-        opts
-      ) do
-    value_column = CQLUtils.type_to_db_column_name(endpoint.value_type)
-    keyspace = Realm.keyspace_name(realm_name)
-
-    {timestamp_ms, timestamp_submillis} = DateTimeMs.split_submillis(timestamp)
-
-    # TODO: :reception_timestamp_submillis is just a place holder right now
-    %Endpoint{endpoint_id: endpoint_id} = endpoint
-
-    interface_storage_attributes = %{
-      value_column => to_db_friendly_type(value),
-      device_id: device_id,
-      interface_id: interface_descriptor.interface_id,
-      endpoint_id: endpoint_id,
-      path: path,
-      reception_timestamp: timestamp_ms,
-      reception_timestamp_submillis: timestamp_submillis
-    }
-
-    opts = [
-      prefix: keyspace,
-      ttl: opts[:ttl],
-      consistency: Consistency.device_info(:write)
-    ]
-
-    {1, _} =
-      Repo.insert_all(interface_descriptor.storage, [interface_storage_attributes], opts)
-
-    :ok
-  end
-
-  # TODO Copy&pasted from data updater plant, make it a library
-  def insert_value_into_db(
-        realm_name,
-        device_id,
-        %InterfaceDescriptor{storage_type: :multi_interface_individual_datastream_dbtable} =
-          interface_descriptor,
-        endpoint,
-        path,
-        value,
-        timestamp,
-        opts
-      ) do
-    value_column = CQLUtils.type_to_db_column_name(endpoint.value_type)
-    keyspace = Realm.keyspace_name(realm_name)
-    {timestamp_ms, timestamp_submillis} = DateTimeMs.split_submillis(timestamp)
-
-    attributes = %{
-      value_column => to_db_friendly_type(value),
-      device_id: device_id,
-      interface_id: interface_descriptor.interface_id,
-      endpoint_id: endpoint.endpoint_id,
-      path: path,
-      value_timestamp: timestamp_ms,
-      reception_timestamp: timestamp_ms,
-      reception_timestamp_submillis: timestamp_submillis
-    }
-
-    opts = [
-      prefix: keyspace,
-      ttl: opts[:ttl],
-      consistency: Consistency.time_series(:write, endpoint)
-    ]
-
-    {1, _} = Repo.insert_all(interface_descriptor.storage, [attributes], opts)
-
-    :ok
-  end
-
-  # TODO Copy&pasted from data updater plant, make it a library
-  def insert_value_into_db(
-        realm_name,
-        device_id,
-        %InterfaceDescriptor{storage_type: :one_object_datastream_dbtable} = interface_descriptor,
-        mapping,
-        path,
-        value,
-        timestamp,
-        opts
-      ) do
-    keyspace = Realm.keyspace_name(realm_name)
-    interface_id = interface_descriptor.interface_id
-
-    endpoint_rows =
-      from(Endpoint,
-        where: [interface_id: ^interface_id],
-        select: [:endpoint, :value_type]
-      )
-      |> Repo.all(prefix: keyspace, consistency: Consistency.domain_model(:read))
-
-    explicit_timestamp? = do_interface_has_explicit_timestamp?(keyspace, interface_id)
-
-    column_meta =
-      endpoint_rows
-      |> Map.new(fn endpoint ->
-        endpoint_name = endpoint.endpoint |> String.split("/") |> List.last()
-        column_name = CQLUtils.endpoint_to_db_column_name(endpoint_name)
-        {endpoint_name, %{name: column_name, type: endpoint.value_type}}
-      end)
-
-    base_attributes = %{
-      device_id: device_id,
-      path: path
-    }
-
-    timestamp_attributes = timestamp_attributes(explicit_timestamp?, timestamp)
-    value_attributes = value_attributes(column_meta, value)
-
-    object_datastream =
-      base_attributes
-      |> Map.merge(timestamp_attributes)
-      |> Map.merge(value_attributes)
-
-    ttl = Keyword.get(opts, :ttl)
-
-    opts = [
-      prefix: keyspace,
-      ttl: ttl,
-      returning: false,
-      consistency: Consistency.time_series(:write, mapping)
-    ]
-
-    Repo.insert_all(interface_descriptor.storage, [object_datastream], opts)
-
-    :ok
-  end
-
-  defp timestamp_attributes(true = _explicit_timestamp?, timestamp) do
-    {timestamp, submillis} =
-      Astarte.AppEngine.API.DateTime.split_submillis(timestamp)
-
-    %{
-      value_timestamp: timestamp,
-      reception_timestamp: timestamp,
-      reception_timestamp_submillis: submillis
-    }
-  end
-
-  defp timestamp_attributes(_nil_or_false_explicit_timestamp?, timestamp) do
-    {timestamp, submillis} =
-      Astarte.AppEngine.API.DateTime.split_submillis(timestamp)
-
-    %{reception_timestamp: timestamp, reception_timestamp_submillis: submillis}
-  end
-
-  defp value_attributes(column_meta, value) do
-    value =
-      value
-      |> Enum.flat_map(fn {key, value} ->
-        # filter map
-        case Map.fetch(column_meta, key) do
-          {:ok, meta} ->
-            %{name: name, type: type} = meta
-            data = %{type: type, value: value}
-            [{name, data}]
-
-          :error ->
-            Logger.warning("Unexpected object key #{inspect(key)} with value #{inspect(value)}.")
-
-            []
-        end
-      end)
-
-    value
-    |> Map.new(fn {column, data} -> {column, data.value} end)
-  end
-
-  # TODO Copy&pasted from data updater plant, make it a library
-  defp to_db_friendly_type(array) when is_list(array) do
-    # If we have an array, we convert its elements to a db friendly type
-    Enum.map(array, &to_db_friendly_type/1)
-  end
-
-  defp to_db_friendly_type(%DateTime{} = datetime) do
-    DateTime.to_unix(datetime, :millisecond)
-  end
-
-  defp to_db_friendly_type(value) do
-    value
   end
 
   @device_status_columns_without_device_id [
