+++ libs/astarte_data_access/lib/astarte_data_access/device.ex
@@ -21,10 +21,15 @@ defmodule Astarte.DataAccess.Device do
   This module provides functions to fetch and manipulate device information in Astarte Data Access.
   """
   require Logger
+  alias Astarte.Core.CQLUtils
   alias Astarte.Core.Device, as: DeviceCore
+  alias Astarte.Core.InterfaceDescriptor
+  alias Astarte.Core.Mapping
   alias Astarte.DataAccess.Consistency
+  alias Astarte.DataAccess.Device.InsertContext
   alias Astarte.DataAccess.Device.UnconfirmedDevice
   alias Astarte.DataAccess.Devices.Device
+  alias Astarte.DataAccess.Realms.Endpoint
   alias Astarte.DataAccess.Realms.Realm
   alias Astarte.DataAccess.Repo
 
@@ -214,5 +219,287 @@ defmodule Astarte.DataAccess.Device do
       consistency: consistency,
       error: :device_not_found
     )
+  end
+
+  def insert_value_into_db(
+        %{
+          interface_descriptor: %InterfaceDescriptor{
+            storage_type: :multi_interface_individual_properties_dbtable
+          },
+          mapping: %Mapping{allow_unset: true},
+          value: nil
+        } = context
+      ) do
+    %InsertContext{
+      realm: realm,
+      device_id: device_id,
+      interface_descriptor: interface_descriptor,
+      mapping: mapping,
+      path: path,
+      opts: opts
+    } = context
+
+    %InterfaceDescriptor{storage: storage, interface_id: interface_id} = interface_descriptor
+    %Mapping{endpoint_id: endpoint_id} = mapping
+    keyspace = Realm.keyspace_name(realm)
+
+    _ =
+      remove_property_row(keyspace, storage, device_id, interface_id, endpoint_id, path, opts)
+
+    :ok
+  end
+
+  def insert_value_into_db(
+        %{
+          interface_descriptor: %InterfaceDescriptor{
+            storage_type: :multi_interface_individual_properties_dbtable
+          },
+          value: nil
+        } = context
+      ) do
+    %InsertContext{
+      realm: realm,
+      device_id: device_id
+    } = context
+
+    _ =
+      Logger.warning(
+        "Device #{inspect(device_id)} in realm #{realm} tried to unset an unsettable property.",
+        tag: :unset_not_allowed
+      )
+
+    {:error, :unset_not_allowed}
+  end
+
+  def insert_value_into_db(
+        %{
+          interface_descriptor: %InterfaceDescriptor{
+            storage_type: :multi_interface_individual_properties_dbtable
+          }
+        } = context
+      ) do
+    %InsertContext{
+      realm: realm,
+      device_id: device_id,
+      interface_descriptor: interface_descriptor,
+      mapping: mapping,
+      path: path,
+      value: value,
+      reception_timestamp: reception_timestamp
+    } = context
+
+    %InterfaceDescriptor{interface_id: interface_id, storage: storage} = interface_descriptor
+    %Mapping{endpoint_id: endpoint_id, value_type: value_type} = mapping
+    keyspace_name = Realm.keyspace_name(realm)
+    timestamp = div(reception_timestamp, 10_000)
+    reception_timestamp_submillis = rem(reception_timestamp, 10_000)
+    column_name = CQLUtils.type_to_db_column_name(value_type)
+    db_value = to_db_friendly_type(value)
+
+    # TODO: :reception_timestamp_submillis is just a place holder right now
+    insert_value = %{
+      "device_id" => device_id,
+      "interface_id" => interface_id,
+      "endpoint_id" => endpoint_id,
+      "path" => path,
+      "reception_timestamp" => timestamp,
+      "reception_timestamp_submillis" => reception_timestamp_submillis,
+      column_name => db_value
+    }
+
+    insert_opts = [
+      prefix: keyspace_name,
+      consistency: Consistency.device_info(:write)
+    ]
+
+    _ = Repo.insert_all(storage, [insert_value], insert_opts)
+    :ok
+  end
+
+  def insert_value_into_db(
+        %{
+          interface_descriptor: %InterfaceDescriptor{
+            storage_type: :multi_interface_individual_datastream_dbtable
+          }
+        } = context
+      ) do
+    %InsertContext{
+      realm: realm,
+      device_id: device_id,
+      interface_descriptor: interface_descriptor,
+      mapping: mapping,
+      path: path,
+      value: value,
+      value_timestamp: value_timestamp,
+      reception_timestamp: reception_timestamp,
+      opts: opts
+    } = context
+
+    %InterfaceDescriptor{interface_id: interface_id, storage: storage} = interface_descriptor
+    %Mapping{endpoint_id: endpoint_id, value_type: value_type} = mapping
+    keyspace_name = Realm.keyspace_name(realm)
+    timestamp = div(reception_timestamp, 10_000)
+    reception_timestamp_submillis = rem(reception_timestamp, 10_000)
+    column_name = CQLUtils.type_to_db_column_name(value_type)
+    db_value = to_db_friendly_type(value)
+
+    # TODO: use received value_timestamp when needed
+    # TODO: :reception_timestamp_submillis is just a place holder right now
+    insert_value = %{
+      "device_id" => device_id,
+      "interface_id" => interface_id,
+      "endpoint_id" => endpoint_id,
+      "path" => path,
+      "value_timestamp" => value_timestamp,
+      "reception_timestamp" => timestamp,
+      "reception_timestamp_submillis" => reception_timestamp_submillis,
+      column_name => db_value
+    }
+
+    insert_opts = [
+      prefix: keyspace_name,
+      consistency: Consistency.time_series(:write, mapping)
+    ]
+
+    _ = Repo.insert_all(storage, [insert_value], Keyword.merge(opts, insert_opts))
+
+    :ok
+  end
+
+  def insert_value_into_db(
+        %{
+          interface_descriptor: %InterfaceDescriptor{storage_type: :one_object_datastream_dbtable}
+        } = context
+      ) do
+    %InsertContext{
+      realm: realm,
+      device_id: device_id,
+      interface_descriptor: interface_descriptor,
+      mapping: mapping,
+      path: path,
+      value: value,
+      value_timestamp: value_timestamp,
+      reception_timestamp: reception_timestamp,
+      opts: opts
+    } = context
+
+    %InterfaceDescriptor{interface_id: interface_id, storage: storage} = interface_descriptor
+
+    keyspace_name = Realm.keyspace_name(realm)
+    timestamp = div(reception_timestamp, 10_000)
+    reception_timestamp_submillis = rem(reception_timestamp, 10_000)
+
+    # TODO: we should cache endpoints by interface_id
+    column_info =
+      Endpoint
+      |> select([:endpoint, :value_type])
+      |> where(interface_id: ^interface_id)
+      |> put_query_prefix(keyspace_name)
+      |> Repo.all(consistency: Consistency.domain_model(:read))
+      |> Map.new(fn endpoint ->
+        value_name = endpoint.endpoint |> String.split("/") |> List.last()
+        column_name = CQLUtils.endpoint_to_db_column_name(value_name)
+        {value_name, column_name}
+      end)
+
+    # TODO: we should also cache explicit_timestamp
+    explicit_timestamp_query =
+      from e in Endpoint,
+        prefix: ^keyspace_name,
+        where: e.interface_id == ^interface_id,
+        select: e.explicit_timestamp,
+        limit: 1
+
+    [explicit_timestamp?] =
+      Repo.all(explicit_timestamp_query, consistency: Consistency.domain_model(:read))
+
+    # TODO: use received value_timestamp when needed
+    # TODO: :reception_timestamp_submillis is just a place holder right now
+    insert_params = %{
+      "device_id" => device_id,
+      "path" => path,
+      "reception_timestamp" => timestamp,
+      "reception_timestamp_submillis" => reception_timestamp_submillis
+    }
+
+    object_value = compute_db_object_entries(column_info, value)
+
+    insert_value = Map.merge(insert_params, object_value)
+
+    insert_value =
+      if explicit_timestamp? do
+        Map.put(insert_value, "value_timestamp", value_timestamp)
+      else
+        insert_value
+      end
+
+    insert_opts = [
+      prefix: keyspace_name,
+      consistency: Consistency.time_series(:write, mapping)
+    ]
+
+    _ = Repo.insert_all(storage, [insert_value], Keyword.merge(opts, insert_opts))
+
+    :ok
+  end
+
+  defp remove_property_row(
+         keyspace,
+         table,
+         device_id,
+         interface_id,
+         endpoint_id,
+         path,
+         opts
+       ) do
+    query =
+      from table,
+        prefix: ^keyspace,
+        where: [
+          device_id: ^device_id,
+          interface_id: ^interface_id,
+          endpoint_id: ^endpoint_id,
+          path: ^path
+        ]
+
+    opts = Keyword.merge(opts, consistency: Consistency.device_info(:write))
+
+    _ = Repo.delete_all(query, opts)
+  end
+
+  defp compute_db_object_entries(column_info, object) do
+    Enum.reduce(object, %{}, fn {object_key, object_value}, acc ->
+      case Map.fetch(column_info, object_key) do
+        {:ok, column_name} ->
+          db_value = to_db_friendly_type(object_value)
+          Map.put(acc, column_name, db_value)
+
+        :error ->
+          _ =
+            Logger.warning(
+              "Unexpected object key #{object_key} with value #{inspect(object_value)}."
+            )
+
+          acc
+      end
+    end)
+  end
+
+  defp to_db_friendly_type(array) when is_list(array) do
+    # If we have an array, we convert its elements to a db friendly type
+    Enum.map(array, &to_db_friendly_type/1)
+  end
+
+  defp to_db_friendly_type(%DateTime{} = datetime) do
+    DateTime.to_unix(datetime, :millisecond)
+  end
+
+  # From Cyanide 2.0, binaries are decoded as %Cyanide.Binary{}
+  defp to_db_friendly_type(%Cyanide.Binary{subtype: _subtype, data: bin}) do
+    bin
+  end
+
+  defp to_db_friendly_type(value) do
+    value
   end
 end
